refactor(stats): decompose StatsOverview component for better maintainability
