+++ src/components/Dashboard/StatsSection/StatsOverview/StatsOverview.component.tsx
@@ -1,12 +1,13 @@
-import { memo, useMemo } from "react";
+import { memo } from "react";
 import { motion } from "framer-motion";
 import { LoadingScreen } from "../../../UI/LoadingScreen/LoadingScreen.component";
 import { LevelProgress } from "./LevelProgress.component";
 import { StatCard } from "./StatCard.component";
-import { FaTrophy, FaFire, FaClock, FaStar } from "react-icons/fa";
-import { useUserStats } from "../hooks/useUserStats";
-import { UserStats } from "../../../../types/stats.types";
 import { BadgesGrid } from "./BadgesGrid/BadgesGrid.component";
+import { ErrorState } from "./components/ErrorState.component";
+import { useStatsCards } from "./hooks/useStatsCards";
+import { UserStats } from "../../../../types/stats.types";
+import { useUserStats } from "../hooks/useUserStats";
 
 type StatsOverviewProps = {
   stats: UserStats;
@@ -25,57 +26,14 @@ const container = {
 
 export const StatsOverview = memo(({ stats, isLoading }: StatsOverviewProps) => {
   const { data: statsData, error } = useUserStats();
-
-  const formatTime = useMemo(() => (minutes: number): string => {
-    const hours = Math.floor(minutes / 60);
-    const mins = minutes % 60;
-    return `${hours}h ${mins}m`;
-  }, []);
-
-  const statsCards = useMemo(() => {
-    if (!statsData) return [];
-    
-    return [
-      {
-        icon: FaTrophy,
-        label: "Ukończone Wyzwania",
-        value: statsData.completedChallenges.toString(),
-        gradient: "from-amber-500 to-orange-500"
-      },
-      {
-        icon: FaFire,
-        label: "Aktualny Streak",
-        value: `${statsData.currentStreak} dni`,
-        subValue: `Najlepszy: ${statsData.bestStreak} dni`,
-        gradient: "from-red-500 to-pink-500"
-      },
-      {
-        icon: FaStar,
-        label: "Średni Wynik",
-        value: `${statsData.averageScore}%`,
-        gradient: "from-indigo-500 to-purple-500"
-      },
-      {
-        icon: FaClock,
-        label: "Czas Nauki",
-        value: formatTime(statsData.totalTimeSpent),
-        gradient: "from-emerald-500 to-teal-500"
-      }
-    ];
-  }, [statsData, formatTime]);
+  const statsCards = useStatsCards(statsData);
 
   if (isLoading) {
     return <LoadingScreen />;
   }
 
   if (error) {
-    return (
-      <div className="p-4 bg-red-500/10 border border-red-500/20 rounded-lg">
-        <p className="text-red-500">
-          Wystąpił błąd podczas ładowania statystyk: {error.message}
-        </p>
-      </div>
-    );
+    return <ErrorState message={`Wystąpił błąd podczas ładowania statystyk: ${error.message}`} />;
   }
 
   if (!statsData) {
